+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Tools/ReleasePlan/ProductOnboardingTool.cs
@@ -4,30 +4,32 @@
 using System.CommandLine;
 using System.ComponentModel;
 using Azure.Sdk.Tools.Cli.Commands;
+using Azure.Sdk.Tools.Cli.Helpers;
 using Azure.Sdk.Tools.Cli.Models;
 using Azure.Sdk.Tools.Cli.Models.AzureDevOps;
-using Azure.Sdk.Tools.Cli.Models.Responses.ProductOnboarding;
+using Azure.Sdk.Tools.Cli.Models.Responses.ReleasePlan;
 using Azure.Sdk.Tools.Cli.Services;
 using Azure.Sdk.Tools.Cli.Tools.Core;
 using Microsoft.TeamFoundation.Common;
 using ModelContextProtocol.Server;
 
-namespace Azure.Sdk.Tools.Cli.Tools.ProductOnboarding
+namespace Azure.Sdk.Tools.Cli.Tools.ReleasePlan
 {
     [Description("Product Onboarding Tool type that contains tools to connect to Azure DevOps to work with product onboarding work items")]
     [McpServerToolType]
     public partial class ProductOnboardingTool(  // partial class required due to source generated regex
         IDevOpsService devOpsService,
-        ILogger<ProductOnboardingTool> logger
+        ILogger<ProductOnboardingTool> logger,
+        IEnvironmentHelper environmentHelper
     ) : MCPMultiCommandTool
     {
-        public override CommandGroup[] CommandHierarchy { get; set; } = [SharedCommandGroups.ProductOnboarding];
+        public override CommandGroup[] CommandHierarchy { get; set; } = [SharedCommandGroups.ReleasePlan];
 
         // Commands
-        private const string SyncProductOnboardingCommandName = "sync";
+        private const string OnboardProductCommandName = "onboard-product";
 
         // MCP Tool Names
-        private const string SyncProductOnboardingToolName = "azsdk_sync_product_onboarding";
+        private const string OnboardProductToolName = "azsdk_onboard_product";
 
         // Options
         private readonly Option<Guid> productIdOpt = new("--product-id")
@@ -94,7 +96,7 @@ namespace Azure.Sdk.Tools.Cli.Tools.ProductOnboarding
 
         protected override List<Command> GetCommands() =>
         [
-            new McpCommand(SyncProductOnboardingCommandName, "Sync product onboarding status", SyncProductOnboardingToolName)
+            new McpCommand(OnboardProductCommandName, "Sync product onboarding status", OnboardProductToolName)
             {
                 productIdOpt,
                 productNameOpt,
@@ -115,7 +117,7 @@ namespace Azure.Sdk.Tools.Cli.Tools.ProductOnboarding
             var command = commandParser.CommandResult.Command.Name;
             switch (command)
             {
-                case SyncProductOnboardingCommandName:
+                case OnboardProductCommandName:
                     var productId = commandParser.GetValue(productIdOpt);
                     var productName = commandParser.GetValue(productNameOpt);
                     var productType = commandParser.GetValue(productTypeOpt);
@@ -126,7 +128,7 @@ namespace Azure.Sdk.Tools.Cli.Tools.ProductOnboarding
                     var dataPlane = commandParser.GetValue(dataPlaneOpt);
                     var managementPlane = commandParser.GetValue(managementPlaneOpt);
                     var submitter = commandParser.GetValue(submitterOpt);
-                    return await SyncProductOnboarding(
+                    return await OnboardProduct(
                         productId,
                         productName,
                         productType,
@@ -145,8 +147,10 @@ namespace Azure.Sdk.Tools.Cli.Tools.ProductOnboarding
             }
         }
 
-        [McpServerTool(Name = SyncProductOnboardingToolName), Description("Create or update a product onboarding work item.")]
-        public async Task<ProductOnboardingResponse> SyncProductOnboarding(
+        public static string TestEnvVarName { get; } = "AZSDKTOOLS_AGENT_TESTING";
+
+        [McpServerTool(Name = OnboardProductToolName), Description("Create or update a product onboarding work item.")]
+        public async Task<ProductOnboardingResponse> OnboardProduct(
             Guid productId,
             string productName,
             string productType,
@@ -157,8 +161,7 @@ namespace Azure.Sdk.Tools.Cli.Tools.ProductOnboarding
             string dataPlane,
             string managementPlane,
             string submitter,
-            CancellationToken ct,
-            bool isTest = true)
+            CancellationToken ct)
         {
             try
             {
@@ -234,6 +237,12 @@ namespace Azure.Sdk.Tools.Cli.Tools.ProductOnboarding
                     ManagementPlane = mp,
                     Submitter = submitter,
                 };
+
+                var isTest = environmentHelper.GetBooleanVariable(TestEnvVarName, false);
+                if (isTest)
+                {
+                    logger.LogInformation("'{TestEnvVarName}' environment variable is set to true, onboarding test product", TestEnvVarName);
+                }
 
                 ProductOnboardingWorkItem? productOnboarding = await devOpsService.GetProductOnboardingAsync(status.ProductId, status.ServiceId, ct, isTest);
                 if (productOnboarding == null)